Refactor Platform to avoid naming conflicts with python's platform module.
