+++ src/gui/traymenu.py
@@ -1,4 +1,4 @@
-from tools.platform import Platform
+from tools.toolbox import Platform
 
 
 class Tray(Platform):
